color editor 70%
git-svn-id: http://192.168.0.243:4435/svn/bee/trunk@786 aaee961a-53e7-2249-bd01-b2a8f0c4ce08

+++ src/com/butent/bee/server/modules/ParamHolderBean.java
@@ -182,7 +182,7 @@ public class ParamHolderBean {
         ? parameter.getText(usr.getCurrentUserId()) : parameter.getText();
   }
 
-  public Integer getTime(String module, String name) {
+  public Long getTime(String module, String name) {
     BeeParameter parameter = getModuleParameter(module, name);
     return parameter.supportsUsers()
         ? parameter.getTime(usr.getCurrentUserId()) : parameter.getTime();
